requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.1
fonttools==4.59.2
idna==3.10
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
stripe==12.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
